refactor: renamed ZonalStats files to snake case
