+++ cbsurge/stats/gdal_raster_source.py
